implemented json_format parameter.
