+++ source/root/usr/lib/logdata-anomaly-miner/aminer/input/SimpleByteStreamLineAtomizerFactory.py
@@ -22,7 +22,7 @@ class SimpleByteStreamLineAtomizerFactory(AtomizerFactory):
     All parsed and unparsed atoms are delivered via two lists of handlers.
     """
 
-    def __init__(self, parsing_model, atom_handler_list, event_handler_list, default_timestamp_paths=None):
+    def __init__(self, parsing_model, atom_handler_list, event_handler_list, default_timestamp_paths=None, eol_sep=b'\n'):
         """
         Create the factory to forward data and events to the given lists for each newly created atomizer.
         @param default_timestamp_paths if not empty list, the value of this timestamp field is extracted from parsed atoms and stored
@@ -35,6 +35,7 @@ class SimpleByteStreamLineAtomizerFactory(AtomizerFactory):
             self.default_timestamp_paths = []
         else:
             self.default_timestamp_paths = default_timestamp_paths
+        self.eol_sep = eol_sep
 
     def get_atomizer_for_resource(self, resource_name):
         """
@@ -43,4 +44,4 @@ class SimpleByteStreamLineAtomizerFactory(AtomizerFactory):
         @return a StreamAtomizer object
         """
         return ByteStreamLineAtomizer(self.parsing_model, self.atom_handler_list, self.event_handler_list, 1 << 16,
-                                      self.default_timestamp_paths)
+                                      self.default_timestamp_paths, self.eol_sep)
