+++ source/root/usr/lib/logdata-anomaly-miner/aminer/schemas/BaseSchema.py
@@ -120,7 +120,8 @@
             'schema': {
                 'verbose': {'type': 'boolean', 'required': False, 'default': False},
                 'multi_source': {'type': 'boolean', 'required': False, 'default': False},
-                'timestamp_paths': {'type': ['string', 'list']}
+                'timestamp_paths': {'type': ['string', 'list']},
+                'eol_sep': {'type': 'string', 'required': False, 'default': '\n'}
             }
         }
 }
